Move to server functions and set up proper tests
